@@ -10,12 +10,12 @@
         }
     ],
     "require": {
-        "php": "^8.0",
-        "nikic/php-parser": "^4.18"
+        "php": "^8.2",
+        "nikic/php-parser": "^4.18",
+        "symfony/console": "^6.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^10.5",
-        "symfony/console": "^6.4"
+        "phpunit/phpunit": "^10.5"
     },
     "autoload": {
         "psr-4": {
