@@ -9,13 +9,20 @@ use CodeIntel\Parser\UsageVisitor;
 use PhpParser\Error;
 use PhpParser\ParserFactory;
 use PhpParser\NodeTraverser;
+use PhpParser\Parser;
 
 /**
- * Finds usages of symbols in the indexed codebase
+ * Finds usages of symbols in the indexed codebase using AST analysis
+ * 
+ * This class coordinates the symbol search process by:
+ * 1. Parsing PHP files using nikic/php-parser
+ * 2. Analyzing AST nodes with UsageVisitor
+ * 3. Scoring confidence levels with ConfidenceScorer
+ * 4. Returning structured usage data for Claude Code
  */
 class UsageFinder
 {
-    private $parser;
+    private Parser $parser;
     private ConfidenceScorer $scorer;
     
     public function __construct(
@@ -38,7 +45,8 @@ class UsageFinder
         
         foreach ($indexedFiles as $filePath) {
             $usages = $this->findUsagesInFile($symbolName, $filePath);
-            $allUsages = array_merge($allUsages, $usages);
+            // Use array_push with spread operator for better performance
+            array_push($allUsages, ...$usages);
         }
         
         return $allUsages;
@@ -57,22 +65,26 @@ class UsageFinder
                 return [];
             }
             
-            $visitor = new UsageVisitor($symbolName, $filePath);
+            $visitor = new UsageVisitor($symbolName, $filePath, $code);
+            
+            // Create fresh traverser for each file to avoid visitor conflicts
             $traverser = new NodeTraverser();
             $traverser->addVisitor($visitor);
             $traverser->traverse($ast);
             
             $usages = $visitor->getUsages();
             
-            // Apply confidence scoring
+            // Apply confidence scoring in place for better memory usage
             foreach ($usages as &$usage) {
                 $usage['confidence'] = $this->scorer->score($usage['code']);
             }
+            unset($usage); // Break reference to last element
             
             return $usages;
             
         } catch (Error $e) {
-            // Parsing error - skip this file
+            // Parsing error - skip this file silently for now
+            // In production, this could be logged
             return [];
         }
     }
